WIP oauth with facebook functional struggeling with connecting the two auth methods together
