+++ package.json
@@ -3,10 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "lodash": "^4.17.15",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
-    "react-facebook-auth": "^1.4.0",
-    "react-facebook-login": "^4.1.1",
     "react-hook-form": "^3.23.1",
     "react-redux": "^7.1.1",
     "react-scripts": "3.1.1",
