+++ src/index.js
@@ -6,8 +6,8 @@ import * as serviceWorker from "./serviceWorker";
 import "semantic-ui-css/semantic.min.css";
 import configureStore from "./state/store/configureStore";
 import { verifyCredentials } from "./state/actions/reduxTokenAuthConfig";
-// import { reduxOauthConfig } from './state/actions/oauthActions'
 import { initialize } from 'redux-oauth';
+
 const reduxOauthConfig = {
   backend: {
     apiUrl: 'http://localhost:3000/api',
